feat: 존재하지 않는 태그 삭제 예외 처리 추가

+++ Domain-Module/src/main/kotlin/com/bamyanggang/domainmodule/domain/tag/service/TagRemover.kt
@@ -1,5 +1,6 @@
 package com.bamyanggang.domainmodule.domain.tag.service
 
+import com.bamyanggang.domainmodule.domain.tag.exception.TagException
 import com.bamyanggang.domainmodule.domain.tag.repository.TagRepository
 import org.springframework.stereotype.Service
 import java.util.*
@@ -9,6 +10,9 @@ class TagRemover(
     private val tagRepository: TagRepository
 ) {
     fun removeTag(tagId: UUID) {
+        if (tagRepository.isExistById(tagId)) {
+            throw TagException.NotFoundTag()
+        }
         tagRepository.deleteByTagId(tagId)
     }
 }
